Title: 최대 숫자 만들기, Time: 405ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240223/최대 숫자 만들기/make-biggest-num.cpp
@@ -4,19 +4,19 @@
 
 using namespace std;
 
-bool cmp(int a, int b)
+bool cmp(long long a, long long b)
 {
-    return stoi(to_string(a)+to_string(b)) > stoi(to_string(b)+to_string(a));
+    return stoll(to_string(a)+to_string(b)) > stoll(to_string(b)+to_string(a));
 }
 
 int main() {
     int n;
     scanf("%d",&n);
 
-    int arr[n];
-    for (int i=0;i<n;i++) scanf("%d",&arr[i]);
+    long long arr[n];
+    for (int i=0;i<n;i++) scanf("%lld",&arr[i]);
     sort(arr,arr+n,cmp);
 
-    for (int i=0;i<n;i++) printf("%d",arr[i]);
+    for (int i=0;i<n;i++) printf("%lld",arr[i]);
     return 0;
 }
